[test][e2e] external transfer for double wallet (#472)

+++ scripts/lrw.sh
@@ -42,8 +42,8 @@ COMPOSE_YAML=docker/docker-compose.yaml
 COMPOSE_DEV_YAML=docker/dev.docker-compose.yaml
 COMPOSE_DEBUG_YAML=docker/debug.docker-compose.yaml
 COMPOSE_E2E_YAML=docker/e2e.docker-compose.yaml
-COMPOSE_DOUBLE_YAML=docker/double.docker-compose.yaml
-COMPOSE_E2E_DOUBLE_YAML=docker/e2e.double.docker-compose.yaml
+COMPOSE_E2E_BLIND=docker/e2e.blind.docker-compose.yaml
+COMPOSE_E2E_TEST=docker/e2e.test.docker-compose.yaml
 COMPOSE_STATIC_YAML=docker/static.docker-compose.yaml
 PG_VOLUME=pg-data
 PG_VOLUME_DOUBLE=pg-data-2
@@ -136,29 +136,30 @@ run() {
 
 build() {
   local port=${1:-8080}
-  local is_helm=${2:-false}
+  local build_mode=${2:-dev}
   echo "production mode with gw port ${port}"
-  echo "helm chart mode is ${is_helm}"
+  echo "build mode is ${build_mode}"
 
-  warn "removing old build artifacts..."
-  rm -fr "${project_dir}/gateway/tmp/frontend/"
-
-
-  info "build the frontend to a static production package"
-  mkdir -p "${project_dir}/gateway/tmp/frontend/"
-  warn "done"
-  info "running docker to compile frontend..."
-  docker build -t reference-wallet-frontend-build -f "${project_dir}/frontend/Dockerfile" "${project_dir}/frontend/" || fail 'frontend container build failed!'
-  docker create --name tmp_reference_frontend reference-wallet-frontend-build || fail 'frontend compilation failed!'
-  docker cp tmp_reference_frontend:/app/build/. "${project_dir}/gateway/tmp/frontend/" || fail 'frontend copy artifacts failed!'
-  docker rm tmp_reference_frontend
-  info "frontend build completed"
-
-  # build all the service images using compose
-  if [ "$is_helm" == false ]; then
-    docker-compose -f ${COMPOSE_YAML} -f ${COMPOSE_DEV_YAML} build || fail 'docker-compose build failed!'
-  else
+  if [ "$build_mode" = "helm" ]; then
     docker-compose -f ${COMPOSE_YAML} -f ${COMPOSE_STATIC_YAML} build  || fail 'docker-compose build failed!'
+  elif [ "$build_mode" = "e2e" ]; then
+    docker-compose -f ${COMPOSE_YAML} build  || fail 'docker-compose build failed!'
+  elif [ "$build_mode" = "e2e-blind" ]; then
+    docker-compose -f ${COMPOSE_YAML} -f ${COMPOSE_E2E_BLIND} build  || fail 'docker-compose build failed!'
+  else
+    warn "removing old build artifacts..."
+    rm -fr "${project_dir}/gateway/tmp/frontend/"
+
+    info "build the frontend to a static production package"
+    mkdir -p "${project_dir}/gateway/tmp/frontend/"
+    warn "done"
+    info "running docker to compile frontend..."
+    docker build -t reference-wallet-frontend-build -f "${project_dir}/frontend/Dockerfile" "${project_dir}/frontend/" || fail 'frontend container build failed!'
+    docker create --name tmp_reference_frontend reference-wallet-frontend-build || fail 'frontend compilation failed!'
+    docker cp tmp_reference_frontend:/app/build/. "${project_dir}/gateway/tmp/frontend/" || fail 'frontend copy artifacts failed!'
+    docker rm tmp_reference_frontend
+    info "frontend build completed"
+    docker-compose -f ${COMPOSE_YAML} -f ${COMPOSE_DEV_YAML} build || fail 'docker-compose build failed!'
   fi
 
 }
@@ -192,12 +193,13 @@ get_port() {
 }
 
 e2e() {
-  
   local single_double=${1}
   local up_down=$2
   local upgrade=${3:-false} # use when upgrading single to double
   composes="${COMPOSE_YAML} -f ${COMPOSE_E2E_YAML}"
-  volumes="docker_${PG_VOLUME}"
+  volumes="lrw_${PG_VOLUME}"
+
+  export COMPOSE_PROJECT_NAME=""
 
   if [ "$single_double" = "single" ]; then
     if [ "$upgrade" == true ]; then
@@ -206,8 +208,7 @@ e2e() {
     fi
     echo "Operating on single wallet instance"
   elif [ "$single_double" = "double" ]; then
-    double_composes="$composes -f ${COMPOSE_DOUBLE_YAML} -f ${COMPOSE_E2E_DOUBLE_YAML}"
-    volumes="$volumes docker_${PG_VOLUME_DOUBLE}"
+    volumes="$volumes lrw2_${PG_VOLUME_DOUBLE}"
     echo "Operating on double wallet instance"
   else
     echo 'Must specify valid count `single` or `double`'
@@ -215,30 +216,52 @@ e2e() {
 
   if [ "$up_down" = "up" ]; then
     if [ "$upgrade" == false ]; then
+
+      # set up the test fixtures first
+      docker-compose -f $COMPOSE_E2E_TEST up --detach > /dev/null 2>&1
+
       export GW_PORT=$(get_port)
       ENV_FILE_NAME=".env" PIPENV_PIPFILE=backend/Pipfile PIPENV_DONT_LOAD_ENV=1 GW_PORT=$GW_PORT \
-            LIQUIDITY_SERVICE_PORT=5000 LIQUIDITY_SERVICE_HOST="liquidity" \
             pipenv run python3 scripts/set_env.py > /dev/null 2>&1
       export VASP_ADDR_1=$(source backend/.env && echo $VASP_ADDR)
       echo "export GW_PORT_1=$GW_PORT"
       echo "export VASP_ADDR_1=$VASP_ADDR_1"
-      docker-compose -f $composes up --detach > /dev/null 2>&1
+      docker-compose -p lrw -f $composes up --detach > /dev/null 2>&1
+      docker network connect lrw_default test-runner
     fi
     if [ "$single_double" = "double" ]; then
       export GW_PORT_2=$(get_port)
       ENV_FILE_NAME=".env-2" PIPENV_PIPFILE=backend/Pipfile PIPENV_DONT_LOAD_ENV=1 GW_PORT=$GW_PORT_2 \
-            LIQUIDITY_SERVICE_PORT=5000 LIQUIDITY_SERVICE_HOST="liquidity-2" \
             pipenv run python3 scripts/set_env.py > /dev/null 2>&1
       export VASP_ADDR_2=$(source backend/.env-2 && echo $VASP_ADDR)
       echo "export GW_PORT_2=$GW_PORT_2"
       echo "export VASP_ADDR_2=$VASP_ADDR_2"
-      docker-compose -f $double_composes up --detach > /dev/null 2>&1
+
+      # !!!hacking the env file!!!
+      # saves .env to temp and replaces with .env-2
+      tmp_backend_env=$(mktemp)
+      tmp_liquidity_env=$(mktemp)
+      backend_env="${project_dir}/backend/.env"
+      liquidity_env="${project_dir}/liquidity/.env"
+      backend_env_2="${project_dir}/backend/.env-2"
+      liquidity_env_2="${project_dir}/liquidity/.env-2"
+      cp $backend_env $tmp_backend_env
+      cp $liquidity_env $tmp_liquidity_env
+      cp $backend_env_2 $backend_env
+      cp $liquidity_env_2 $liquidity_env
+
+      export GW_PORT=$GW_PORT_2
+      docker-compose -p lrw2 -f $composes up --detach > /dev/null 2>&1
+      docker network connect lrw2_default test-runner
+
+      cp $tmp_backend_env $backend_env
+      cp $tmp_liquidity_env $liquidity_env
     fi
   elif [ "$up_down" = "down" ]; then
-    if [ "$single_double" = "single" ]; then
-      docker-compose -f $composes down
-    else
-      docker-compose -f $double_composes down
+    docker-compose -f $COMPOSE_E2E_TEST down
+    docker-compose -p lrw -f $composes down
+    if [ "$single_double" = "double" ]; then
+      docker-compose -p lrw2 -f $composes down
     fi
     echo "Purging $volumes"
     docker volume rm -f $volumes > /dev/null 2>&1
@@ -303,6 +326,7 @@ watch_test() {
 }
 
 setup_environment() {
+  local skip_build=${1:-false}
   if ! command -v pipenv &> /dev/null
   then
     ec "Installing pipenv"
@@ -319,14 +343,19 @@ setup_environment() {
   info "***Installing liquidity dependencies***"
   sh -c "cd liquidity && pipenv install --dev"
 
-  info "***Installing frontend dependencies***"
-  sh -c "cd frontend && yarn"
-
   info "***Setting up environment .env files***"
   PIPENV_PIPFILE=backend/Pipfile pipenv run python3 scripts/set_env.py
 
   info "***Setting up docker-compose project name***"
   cp .env.example .env
+
+  if [ "$skip_build" == true ]; then
+    echo "Skipping build"
+    exit 0
+  fi
+
+  info "***Installing frontend dependencies***"
+  sh -c "cd frontend && yarn"
 
   info "***Building docker images***"
   build
